+++ lib/assets/javascripts/new-dashboard/components/MapCard/CondensedMapHeader.vue
@@ -27,7 +27,7 @@
             {{ $t(`MapListHeader.lastModified`) }}
           </span>
         </div>
-        <div class="cell cell--large u-txt-right u-p0" @click="changeOrder('mapviews')">
+        <div class="cell cell--large u-txt-right u-p--0 " @click="changeOrder('mapviews')">
           <span class="text is-small is-txtSoftGrey"
                 :class="{
                   'is-active': isOrderApplied('mapviews'),